+++ src/test/utility_test.rs
@@ -6,6 +6,7 @@ use ndarray::prelude::*;
 mod kernel_pca_test;
 mod label_encoding_test;
 mod linear_discriminant_analysis_test;
+mod normalize_test;
 mod principal_component_analysis_test;
 mod t_sne_test;
 mod train_test_split_test;